+++ test/bootstrap_fields_test.rb
@@ -416,4 +416,14 @@ class BootstrapFieldsTest < ActionView::TestCase
       end
     end
   end
+
+  test "can have a floating label" do
+    expected = <<~HTML
+      <div class="mb-3 form-floating">
+        <input class="form-control" id="user_email" name="user[email]" type="text" value="steve@example.com" placeholder="Email" />
+        <label class="form-label required" for="user_email">Email</label>
+      </div>
+    HTML
+    assert_equivalent_xml expected, @builder.text_field(:email, floating: true)
+  end
 end